refactor(login-fragment): move shared pref code to specialized class #24
